+++ modules/logging/ui/src/sap/commerce/toolset/logging/ui/LoggersSplitView.kt
@@ -26,6 +26,7 @@ import com.intellij.ui.components.JBScrollPane
 import com.intellij.util.asSafely
 import kotlinx.coroutines.CoroutineScope
 import kotlinx.coroutines.launch
+import sap.commerce.toolset.exec.settings.state.ExecConnectionScope
 import sap.commerce.toolset.hac.exec.HacExecConnectionService
 import sap.commerce.toolset.hac.exec.settings.event.HacConnectionSettingsListener
 import sap.commerce.toolset.hac.exec.settings.state.HacConnectionSettingsState
@@ -69,8 +70,10 @@ class LoggersSplitView(
 
         with(project.messageBus.connect(this)) {
             subscribe(HacConnectionSettingsListener.TOPIC, object : HacConnectionSettingsListener {
-                override fun onActiveConnectionChanged(connection: HacConnectionSettingsState) = updateTree(connection)
-                override fun onModified(connection: HacConnectionSettingsState) = tree.update()
+                override fun onActive(connection: HacConnectionSettingsState) = updateTree(connection)
+                override fun onSave(settings: Map<ExecConnectionScope, List<HacConnectionSettingsState>>) = settings.values
+                    .flatten()
+                    .forEach { updateTree(it) }
             })
 
             subscribe(CxLoggersStateListener.TOPIC, object : CxLoggersStateListener {