+++ CHANGELOG.md
@@ -22,6 +22,7 @@
 
 ### `hAC` enhancements
 - Use own configuration pane for hAC-related connection settings [#1543](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1543)
+- Extracted HacHttpClient Cache and adjusted notifications [#1557](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1557)
 
 ### `Solr` enhancements
 - Use own configuration pane for Solr-related connection settings [#1541](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/1541)